Rename to agreed schema.
