+++ graphql-api.cabal
@@ -36,12 +36,12 @@ library
     , QuickCheck
     , text
   exposed-modules:
-      GraphQL.Input
-      GraphQL.Output
-      GraphQL.Schema
-      GraphQL.TypedApi
-      GraphQL.TypedSchema
-      GraphQL.Validation
+      GraphQL.API
+      GraphQL.Internal.Input
+      GraphQL.Internal.Output
+      GraphQL.Internal.Schema
+      GraphQL.Internal.Validation
+      GraphQL.Server
       GraphQL.Value
   default-language: Haskell2010
