+++ tests/TypeApiTests.hs
@@ -6,13 +6,13 @@ import Protolude hiding (Enum)
 import Test.Tasty (TestTree)
 import Test.Tasty.Hspec (testSpec, describe, it, shouldBe)
 
-import GraphQL.TypedSchema
+import GraphQL.API
   ( Object
   , Field
   , Argument
   , (:>)
   )
-import GraphQL.TypedApi
+import GraphQL.Server
   ( Handler
   , QueryError(..)
   , buildResolver
